게시물 작성시 DB에 데이터 추가

+++ src/main/java/com/sbs/board/bundedContext/article/dto/Article.java
@@ -1,11 +1,11 @@
 package com.sbs.board.bundedContext.article.dto;
 
 public class Article {
-  public int id;
+  public long id;
   public String title;
   public String content;
 
-  public Article(int id, String title, String content) {
+  public Article(long id, String title, String content) {
     this.id = id;
     this.title = title;
     this.content = content;
